Fix: Refactored code to make file sizes smaller
