+++ keys.go
@@ -5,9 +5,10 @@ import (
 	"context"
 	"os"
 	"os/exec"
+	"net/http"
+	"strings"
 
 	shell "github.com/ipfs/go-ipfs-api"
-
 )
 
 // This function generates a key using local node and returns it
@@ -71,4 +72,97 @@ func importKey(keyName string, fileName string) error {
 	}
 	fmt.Println(string(stdout))
 	return nil
+}
+
+
+// This function generates a key, exports it to "<keyName>.key" file in current dir, then delete from local keystore.
+// returns newly generated key file to user
+func GetKey(w http.ResponseWriter, r *http.Request) {
+	keyName := "temp" // user input from API or self-generated non-clashing name
+	fmt.Println("Generating key...")
+	key, err := genKey(keyName)
+	if err != nil {
+		writeJSONError(w, "Error in genKey", err)
+		return
+	}
+	
+	fmt.Println("Exporting key...")
+	err = exportKey(keyName)
+	if err != nil {
+		writeJSONError(w, "Error in keyName", err)
+		return
+	}
+	
+	fmt.Println("Deleting key...")
+	err = deleteKey(key.Name) // delete temp key from local node keystore
+	if err != nil {
+		writeJSONError(w, "Error in deleteKey", err)
+		return
+	}
+
+
+	// log.Println("Url Param 'keyName' is: " + string(keyName))
+	w.Header().Set("Content-Disposition", "attachment; filename="+string(keyName))
+	w.Header().Set("Content-Type", "application/octet-stream")
+	w.WriteHeader(http.StatusOK)
+	http.ServeFile(w, r, string(keyName) + ".key") // serve key to user to download
+
+
+	fmt.Println("Deleting exported key...")
+	err = diskDelete(keyName+".key") // delete key from disk
+	if err != nil {
+		panic(err)
+		return
+	}
+
+	// fmt.Println("Deleting exported key a second time...")
+	// err = diskDelete(keyName) // delete key from disk again to force error
+	// if err != nil {
+	// 	writeJSONError(w, "Error in deleteKey", err)
+	// 	return
+	// }
+}
+
+// This function will save a key to node, then delete the uploaded file from disk
+// Returns 200 & key name as confirmation
+func PostKey(w http.ResponseWriter, r *http.Request) {
+	const dir = "KeyStore"
+	FileName, err := saveFile(r, dir, 32 << 10) // grab uploaded .key file
+	if err != nil {
+		writeJSONError(w, "Error in saveFile", err)
+		return
+	}
+	name := strings.Split(FileName, ".")[0]
+
+	fmt.Println("Importing Key...")
+	err = importKey(name, dir+"/"+FileName) //import key to local node keystore
+	if err != nil {
+		writeJSONError(w, "Error in importKey", err)
+		return
+	}
+
+	fmt.Println("Deleting saved key from disk...")
+	err = diskDelete(dir+"/"+FileName) // delete key from disk
+	if err != nil {
+		writeJSONError(w, "Error in deleteKey", err)
+		return
+	}
+	writeJSONSuccess(w, "Success - Saved key", name)
+}
+
+// This function will delete a key from the local node keystore
+func DeleteKey(w http.ResponseWriter, r *http.Request) {
+	keyName, ok := GetParam(r, "keyName")
+	if ok != true {
+		writeJSONError(w, keyName, nil)
+		return
+	}
+
+	fmt.Println("Deleting key...")
+	err := deleteKey(keyName) // delete temp key from local node keystore
+	if err != nil {
+		writeJSONError(w, "Error in deleteKey", err)
+		return
+	}
+	writeJSONSuccess(w, "Success - Deleted key", keyName)
 }
